Drop SetupHelpers.Proxied and move everything into Test.Proxy.
